drivers: modem: fix include paths after driver reorganization
Fix relative include paths after moving standalone modem drivers into
per-driver subdirectories.

Update legacy HL7800, Quectel BG9x, ST87Mxx, u-blox SARA R4, and WNC
M14A2A sources to include shared modem helper headers from the new
directory depth.

Also clean up the ST87Mxx Kconfig select ordering after the move.
